move reject dialog to its own component
