+++ src/pages/admin/AdminPage.vue
@@ -55,33 +55,12 @@
     </q-card>
   </q-page>
 
-  <q-dialog v-model="rejectDialog" persistent>
-    <q-card class="q-pa-md" style="border-radius: 15px; background-color: gray">
-      <q-card-section class="text-center">
-        <div class="text-h6 text-white">
-          Are you sure you want to reject {{ selectedApplicant }}?
-        </div>
-      </q-card-section>
-      <q-card-actions>
-        <q-btn
-          flat
-          label="Cancel"
-          color="primary"
-          @click="rejectDialog = false"
-        />
-        <q-btn
-          flat
-          label="Reject"
-          color="negative"
-          @click="rejectApplication"
-        />
-      </q-card-actions>
-    </q-card>
-  </q-dialog>
+  <RejectDialog ref="rejectDialogRef" />
 </template>
 
 <script setup>
 import { ref } from 'vue';
+import RejectDialog from 'src/components/admin/RejectDialog.vue'; // Import the RejectDialog component
 
 const applications = ref([
   {
@@ -146,18 +125,12 @@ const columns = [
   },
 ];
 
-const rejectDialog = ref(false);
-
-const selectedApplicant = ref('');
+const rejectDialogRef = ref(null);
 
 const showRejectDialog = (row) => {
-  selectedApplicant.value = row.applicant;
-  rejectDialog.value = true;
-};
-
-const rejectApplication = () => {
-  console.log('Application rejected');
-  rejectDialog.value = false;
+  if (rejectDialogRef.value) {
+    rejectDialogRef.value.showRejectDialog(row); // Call showRejectDialog on the ref
+  }
 };
 </script>
